Update version from `1.0.3` to `1.1.0` after clean up of obsolete code Changes made: - Removed obsolete methods in `StubServer`:   - `CreateApiService<TApiService>`   - `CreateApiService(type type>)` - Moved the folder `Services` to the folder `Internal`, including classes - Updated access modifier for classes in the folder `Services` from `public` to `internal`. The following classes are impacted:   - `Results/ApiServiceResult`   - `Results/ApiServiceResult<TModel>`   - `ApiService`   - `ReceivedRequestApiService`   - `SetupResponseApiService` - Change access modifier for class `UriExtensions` from `public` to `internal` and moved the class to the folder `Internal` - Removed interface `IApiResult`
