+++ source/Cympatic.Extensions.Stub.UnitTests/UriExtensionsTests.cs
@@ -1,4 +1,5 @@
-﻿using FluentAssertions;
+﻿using Cympatic.Extensions.Stub.Internal;
+using FluentAssertions;
 using System.Collections.Specialized;
 
 namespace Cympatic.Extensions.Stub.UnitTests;
