+++ source/Cympatic.Extensions.Stub/Cympatic.Extensions.Stub.csproj
@@ -7,7 +7,7 @@
     <LangVersion>latest</LangVersion>
 
     <PackageId>Cympatic.Extensions.Stub</PackageId>
-    <PackageVersion>1.0.3</PackageVersion>
+    <PackageVersion>1.1.0</PackageVersion>
     <Authors>Cympatic</Authors>
     <Company>Cympatic B.V.</Company>
     <Description>A lightweight stub server that mimics the functionality of an external service, commonly used by microservices.</Description>
